requirement.txt
alembic==1.16.4
antiorm==1.2.1
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
DateTime==5.5
db==0.1.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-CKEditor==1.0.0
flask-fontawesome==0.2.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Modals==0.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.10
is-disposable-email==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.3.0
pytz==2025.2
SQLAlchemy==2.0.43
typing_extensions==4.14.1
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
wtforms-validators==1.0.0
zope.interface==7.2
